avatar: refact - service dedicated to avatar upload (AvatarUploadService)
