+++ src/frontend/src/app/avatar/avatar-upload/avatar-upload.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, OnDestroy, OnInit } from '@angular/core';
 import { assets } from 'src/assets/assets';
-import { AvatarService } from '../avatar.service';
+import { AvatarUploadService } from './avatar-upload.service';
 
 @Component({
   selector: 'app-avatar-upload',
@@ -10,21 +10,21 @@ import { AvatarService } from '../avatar.service';
 export class AvatarUploadComponent implements OnInit, OnDestroy {
   @Input() default_src = assets.defaultAvatar;
 
-  constructor(public readonly avatar: AvatarService) {}
+  constructor(public readonly service: AvatarUploadService) {}
 
   ngOnInit(): void {
-    if (this.default_src != this.avatar.default_src)
-      this.avatar.update({ src: this.default_src });
+    if (this.default_src != this.service.default_src)
+      this.service.update({ src: this.default_src });
   }
   ngOnDestroy(): void {
-    this.avatar.clear();
+    this.service.clear();
   }
 
-  processUpload(upload: HTMLInputElement) {
+  processInput(upload: HTMLInputElement) {
     if (upload.files) {
       const image = upload.files[0];
       console.log(`Loaded image: ${image.name}`);
-      this.avatar.update({ file: image });
+      this.service.update({ file: image });
     }
   }
 }
