+++ src/frontend/src/app/avatar/avatar-upload/avatar-upload.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { AvatarUploadService } from './avatar-upload.service';
+
+describe('AvatarUploadService', () => {
+  let service: AvatarUploadService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(AvatarUploadService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
